LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/custom-property-and-allow-discrete-expected.txt


PASS It is possible to transition an unregistered custom property using "transition-behavior: allows-discrete"
